feat: Add endpoint for pr outliers
This endpoints returns outlier pull requests.

To identify outliers using quartiles, we are applying the following logic:

Quartiles divide an ordered data set into four equal parts.

- **Q1 (First Quartile):** The value below which 25% of the data falls.
- **Q2 (Median):** The value that splits the data in half, meaning 50% of the data is below it and 50% is above.
- **Q3 (Third Quartile):** The value below which 75% of the data falls.

The difference between Q3 and Q1 is called the **Interquartile Range (IQR)** and represents the spread of the central data.

We consider a value to be an outlier if it is significantly distant from the rest of the data. A common rule is to define outliers as values that are lower than:

```plaintext
Q1 - 1.5 * IQR
```

or higher then

```plaintext
Q3 + 1.5 * IQR
```